Moved IoT and MQTT operations into AWSProvider class
